requirements.txt
pyttsx3~=2.90
PyAudio~=0.2.11
selenium~=4.1.5
Flask~=3.0.2
Werkzeug~=3.0.1
requests~=2.31.0
pillow~=10.2.0
wikipedia~=1.4.0
Pygments~=2.17.2